# public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: LPT APis
  version: 1.0.0
  title: LPT Frontend API's
basePath: /api/
...
paths:
...
  /booking/{user}:
    post:
      tags:
        - Booking
      summary: Booking
      description: 'booking'
      operationId: booking
      consumes:
        - application-json
      produces:
        - application-json
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
      parameters:
        - name: user
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: time_id
          type: integer
          required: true
        - in: query
          name: name
          type: string
          required: true
        - in: query
          name: phone_number
          type: string
          required: true
        - in: query
          name: date
          type: string
          required: true
...